dotbot/frontend: factorize websocket dotbot update logic
